IMN-370 Move v2converters from catalog-process to models (#295)
